Rename vector2_extension to vector2
