+++ doc/examples/aseprite/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/sprite_animation.dart';
-import 'package:flame/vector2_extension.dart';
+import 'package:flame/vector2.dart';
 import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
