+++ doc/examples/joystick/lib/player.dart
@@ -5,7 +5,7 @@ import 'package:flame/components/component.dart';
 import 'package:flame/components/joystick/joystick_component.dart';
 import 'package:flame/components/joystick/joystick_events.dart';
 import 'package:flame/palette.dart';
-import 'package:flame/vector2_extension.dart';
+import 'package:flame/vector2.dart';
 
 class Player extends Component implements JoystickListener {
   final _whitePaint = BasicPalette.white.paint;
