+++ doc/examples/sprites/lib/main.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flame/components/sprite_component.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
-import 'package:flame/vector2_extension.dart';
+import 'package:flame/vector2.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
