+++ doc/examples/text/lib/main.dart
@@ -7,7 +7,7 @@ import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/palette.dart';
 import 'package:flame/text_config.dart';
-import 'package:flame/vector2_extension.dart';
+import 'package:flame/vector2.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
