+++ lib/anchor.dart
@@ -1,4 +1,4 @@
-import 'vector2_extension.dart';
+import 'vector2.dart';
 
 class Anchor {
   static const Anchor topLeft = Anchor(0.0, 0.0);
