+++ lib/particles/component_particle.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 
 import '../particle.dart';
 import '../components/component.dart';
-import '../vector2_extension.dart';
+import '../vector2.dart';
 
 class ComponentParticle extends Particle {
   final Component component;
