+++ test/components/composed_component_test.dart
@@ -5,7 +5,7 @@ import 'package:flame/components/mixins/has_game_ref.dart';
 import 'package:flame/components/mixins/resizable.dart';
 import 'package:flame/components/mixins/tapable.dart';
 import 'package:flame/game/base_game.dart';
-import 'package:flame/vector2_extension.dart';
+import 'package:flame/vector2.dart';
 import 'package:flutter/gestures.dart';
 import 'package:test/test.dart';
